safari-534.53-branch/LayoutTests/http/tests/security/contentSecurityPolicy/javascript-url-allowed.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.layoutTestController)
  layoutTestController.dumpAsText();
</script>
</head>
<body>
  <iframe src="http://127.0.0.1:8000/security/contentSecurityPolicy/resources/javascript-url.pl?should_run=yes&csp=allow%20*%3B"></iframe>
</body>
</html>
