LayoutTests/http/tests/security/contentSecurityPolicy/javascript-url-allowed-expected.txt

CONSOLE MESSAGE: line 1: Unrecognized Content-Security-Policy directive 'allow'.

ALERT: PASS
